api/requirements.txt
SQLAlchemy==1.3.22
requests==2.25.1
pymysql==1.0.2
tenacity==6.3.1
# tensorflow
numpy
scikit-image
matplotlib
fastapi[all]
pillow